python-service/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Supabase client
supabase==2.0.0
postgrest==0.13.0

# Azure OpenAI (uses standard openai package with Azure config)
openai>=1.12.0

# Text processing
beautifulsoup4==4.12.2
lxml==4.9.3
html2text==2024.2.26

# LangGraph (for future agentic framework)
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
tenacity==8.2.3
numpy>=1.24.0
websockets>=11.0,<13.0  # Required for supabase realtime (compatible version)

# Logging
structlog==23.2.0